fix warnings (use of size_t with %d)

+++ lyaml.c
@@ -98,8 +98,8 @@ static void generate_error_message(struct lua_yaml_loader *loader) {
 
    if (loader->parser.problem_mark.line || loader->parser.problem_mark.column) {
       snprintf(buf, sizeof(buf), ", line: %d, column: %d\n",
-         loader->parser.problem_mark.line + 1,
-         loader->parser.problem_mark.column + 1);
+         (int)loader->parser.problem_mark.line + 1,
+         (int)loader->parser.problem_mark.column + 1);
       luaL_addstring(&b, buf);
    } else {
       luaL_addstring(&b, "\n");
@@ -108,8 +108,8 @@ static void generate_error_message(struct lua_yaml_loader *loader) {
    if (loader->parser.context) {
       snprintf(buf, sizeof(buf), "%s at line: %d, column: %d\n",
          loader->parser.context,
-         loader->parser.context_mark.line + 1,
-         loader->parser.context_mark.column + 1);
+         (int)loader->parser.context_mark.line + 1,
+         (int)loader->parser.context_mark.column + 1);
       luaL_addstring(&b, buf);
    }
